+++ test/e2e/app-dir/app/next.config.js
@@ -5,7 +5,7 @@ module.exports = {
       algorithm: 'sha256',
     },
   },
-  output: 'standalone',
+  // output: 'standalone',
   rewrites: async () => {
     return {
       afterFiles: [
